hwp_picoscope: Move hwp picoscope agent into package
