+++ docker-compose.yml
@@ -30,11 +30,12 @@ services:
   # --------------------------------------------------------------------------
   # CHWP picoscope agent
   # --------------------------------------------------------------------------
-  ocs-hwp-picoscope-agent:
-    image: "ocs-hwp-picoscope-agent"
-    build: ./agents/hwp_picoscope/
-    depends_on:
-      - "socs"
+  # Only works with --privileged, will insist users build image themselves
+  # ocs-hwp-picoscope-agent:
+  #   image: "ocs-hwp-picoscope-agent"
+  #   build: ./docker/hwp_picoscope/
+  #   depends_on:
+  #     - "socs"
 
   # --------------------------------------------------------------------------
   # Wiregrid actuator agent
